requirements.txt
openai==1.3.0
python-dotenv==1.0.0
streamlit==1.37.1
numpy==2.1.1
pandas==2.2.3
setuptools
httpx==0.27.2
requests==2.32.3
